// src/app/constants/enums/home-repair-types.ts
export enum HomeRepairType {
  AcRepair = 'AC Repair',
  RefrigeratorRepair = 'Refrigerator Repair',
  TvRepair = 'TV Repair',
  WashingMachineRepair = 'Washing Machine Repair',
  GeyserRepair = 'Geyser Repair',
  MicrowaveRepair = 'Microwave Repair',
  WaterPurifierRepair = 'RO / Water Purifier Repair',
  InverterBatteryRepair = 'Inverter / Battery Repair',
  ElectricalRepair = 'Electrical Repair',
  Plumbing = 'Plumbing',
  FanCoolerRepair = 'Fan / Cooler Repair',
  MixerGrinderRepair = 'Mixer / Grinder Repair',
  ComputerLaptopRepair = 'Computer / Laptop Repair',
  MobileRepair = 'Mobile Repair',
  FurnitureRepair = 'Furniture Repair',
  OtherHomeApplianceRepair = 'Other Home Appliance Repair',
}

export const HOME_REPAIR_TYPES = Object.values(HomeRepairType);